PYTHON/AnomalyDetection/requirements.txt
MarkupSafe==2.1.5
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
idna==3.7
keras==3.2.1
libclang==18.1.1
markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
protobuf==4.25.3
pygments==2.17.2
requests==2.31.0
rich==13.7.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.2
wrapt==1.16.0
blas==1.0
icc_rt==2022.1.0
intel-openmp==2023.1.0
mkl==2023.1.0
mkl-service==2.4.0
mkl_fft==1.3.8
mkl_random==1.2.4
numpy-base==1.26.4
scipy==1.12.0
tbb==2021.8.0
colorama==0.4.6
tqdm==4.66.2
joblib==1.4.0
threadpoolctl==3.4.0
scikit-learn==1.4.2
tzdata==2024.1
pytz==2024.1
python-dateutil==2.9.0.post0
pandas==2.2.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
pillow==10.3.0
pyparsing==3.1.2
dill==0.3.8
dask==2023.11.0